toolrunner: move schema validation tests to tool_test.go
Validation tests belong in tool_test.go (tests tool.go logic), not
runner_test.go (tests runner orchestration loop). Same package, same
directory, just the right file.
